Extracts click handler functions
This is a 'normal' pattern folks would see within React tutorials to
bind `this` in the constructor function. Another option is to use fat
arrow functions as properties in the class…

    onSelectLeft = () => this.setState({next: "state"});

Though, I'm not sure if this is a real language feature, or just
something that Babel lets you do at the moment. Anyway… I usually do
things this way because it's easier.