Use the updater function version of setState
When using state in a `setState` function, because of the async nature
of react rendering, you should use an updater function rather than
passing an object as it guarantees that you have the current state.

https://reactjs.org/docs/react-component.html#setstate

I tend to use this everywhere, rather than passing an object to
setState, for consistency.